+++ src/js/components/sharedComponents/table/TanStackTable.jsx
@@ -1,10 +1,5 @@
 import React, { useMemo, useState } from 'react';
 // import PropTypes, { shape, oneOf, oneOfType } from 'prop-types';
-import SearchAwardsOperation from 'models/v1/search/SearchAwardsOperation';
-import { isCancel } from 'axios';
-import * as MoneyFormatter from 'helpers/moneyFormatter';
-import * as SearchHelper from 'helpers/searchHelper';
-import { subawardTypeGroups, transactionTypeGroups } from 'dataMapping/search/awardType';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import {
     createColumnHelper,
@@ -17,219 +12,38 @@ import {
     useExpanded
 } from '@tanstack/react-table';
 import { uniqueId } from 'lodash';
-import { performKeywordSearch } from "../../../helpers/keywordHelper";
-import { convertToTitleCase } from "../../../helpers/searchHelper";
-import ReadMore from '../ReadMore';
-import { mockRows } from './mockData';
+import NestedTanStackTable from './NestedTanStackTable';
+import { subAwardDefaultColumns, transactionsDefaultColumns } from '../../../dataMapping/search/tanStackTableColumns';
 
 const TanStackTable = (props) => {
-    const data = mockRows;
-    // const [data, setData] = useState(mockRows);
-    const [subData, setSubData] = useState([]);
-    const [isSubLoading, setIsSubLoading] = useState(false);
+    const [awardId, setAwardId] = useState(null);
     const [expanded, setExpanded] = useState({});
-    const [inFlight, setInFlight] = useState(false);
-    const [error, setError] = useState(false);
     const columnHelper = createColumnHelper();
+    let columnArray = subAwardDefaultColumns;
 
-    let searchRequest = null;
+    if (props.columnType === "transactions") {
+        columnArray = transactionsDefaultColumns;
+    }
 
-    const getSubTable = (awardId, rowId) => {
+    const togglePrimeAward = (primeAwardId, rowId) => {
         if (Object.hasOwn(expanded, rowId)) {
             // assume row is expanded and needs unexpanded
             const newExpanded = { ...expanded };
             delete newExpanded[rowId];
-
+            setAwardId(null);
             setExpanded(newExpanded);
             return null;
         }
-
-        // get subRow Data
-        setIsSubLoading(true);
-        if (searchRequest) {
-            // a request is currently in-flight, cancel it
-            searchRequest.cancel();
-        }
-
-        // get searchParams from state
-        const searchParamsTemp = new SearchAwardsOperation();
-        searchParamsTemp.fromState(props.filters);
-
-        // indicate the request is about to start
-        setInFlight(true);
-        setError(false);
-
-        const requestFields = [
-            "Sub-Award ID",
-            "Sub-Awardee Name",
-            "Sub-Award Amount",
-            "Sub-Award Date",
-            "Sub-Award Description",
-            "Sub-Recipient UEI",
-            "Sub-Recipient Location",
-            "Sub-Award Primary Place of Performance",
-            "Sub-Award Type",
-            "Prime Award ID",
-            "Prime Recipient Name",
-            "Prime Award Recipient UEI",
-            "Awarding Agency",
-            "Awarding Sub Agency",
-            "NAICS",
-            "PSC",
-            "recipient_id",
-            "prime_award_recipient_id"
-        ];
-
-        // needs to be dynamic but for now we will go with defaults
-        searchParamsTemp.awardType =
-            props.isTransactions ? transactionTypeGroups.transaction_contracts : subawardTypeGroups.subcontracts;
-
-        const filters = searchParamsTemp;
-        if (!Object.prototype.hasOwnProperty.call(filters, "selectedAwardIDs")) {
-            filters.selectedAwardIDs = [];
-        }
-        filters.selectedAwardIDs.push(awardId);
-
-        const params = {
-            filters: filters.toParams(),
-            fields: requestFields,
-            page: 1,
-            limit: 100,
-            sort: props.isTransactions ? "Transaction Amount" : "Sub-Award Amount",
-            order: "desc",
-            subawards: true,
-            auditTrail: 'Results Table - Spending by award search'
-        };
-
-        // Set the params needed for download API call
-        if (!params.filters.award_type_codes) {
-            return null;
-        }
-
-        if (props.isTransactions) {
-            params.fields = [
-                "Award ID",
-                "Mod",
-                "Recipient Name",
-                "Transaction Amount",
-                "Action Date",
-                "Transaction Description",
-                "Action Type",
-                "Award Type",
-                "Recipient UEI",
-                "Recipient Location",
-                "Primary Place of Performance",
-                "Awarding Agency",
-                "awarding_agency_id",
-                "Awarding Sub Agency",
-                "NAICS",
-                "PSC",
-                "Assistance Listing"
-            ];
-
-            searchRequest = performKeywordSearch(params);
-        }
-        else {
-            searchRequest = SearchHelper.performSpendingByAwardSearch(params);
-        }
-
-        return searchRequest.promise
-            .then((res) => {
-                const newState = {
-                    inFlight: false
-                };
-
-                const parsedResults = res.data.results.map((result) => ({
-                    ...result,
-                    generated_internal_id: encodeURIComponent(result.generated_internal_id)
-                }));
-
-                // don't clear records if we're appending (not the first page)
-                newState.tableInstance = `${uniqueId()}`;
-                newState.results = parsedResults;
-
-                // request is done
-                searchRequest = null;
-                newState.page = res.data.page_metadata.page;
-                newState.lastPage = !res.data.page_metadata.hasNext;
-                setInFlight(newState.inFlight);
-                setSubData(newState.results);
-
-                // props.setAppliedFilterCompletion(true);
-                // setShowSubRow(rowId);
-                setExpanded((prevState) => ({
-                    ...prevState,
-                    [rowId]: true
-                }));
-            })
-            .catch((err) => {
-                if (!isCancel(err)) {
-                    setInFlight(false);
-                    setError(true);
-                    props.setAppliedFilterCompletion(true);
-                    console.log(err);
-                }
-            });
+        setAwardId(primeAwardId);
+        setExpanded((prevState) => ({
+            ...prevState,
+            [rowId]: true
+        }));
+        return true;
     };
 
-    const pickLocationFormat = (location) => {
-        if (location?.address_line1 && location?.city_name && location?.state_code && location?.zip5) {
-            return `${convertToTitleCase(location.address_line1)}, ${convertToTitleCase(location.city_name)}, ${location.state_code}, ${location.zip5}`;
-        }
-        else if (location?.city_name && location?.state_code && location?.zip5) {
-            return `${convertToTitleCase(location.city_name)}, ${location.state_code}, ${location.zip5}`;
-        }
-        else if (location?.city_name && location?.state_code) {
-            return `${convertToTitleCase(location.city_name)}, ${location.state_code}`;
-        }
-        else if (location?.state_name) {
-            return `${location.state_name}, ${location.location_country_code}`;
-        }
-        else if (location?.city_name && location?.location_country_code) {
-            return `${convertToTitleCase(location.city_name)}, ${location.location_country_code}`;
-        }
-        else if (location?.country_name) {
-            return convertToTitleCase(location.country_name);
-        }
-        else if (location?.location_country_code) {
-            return location.location_country_code;
-        }
-        return '--';
-    };
 
-    const twoVariableFormat = (object, key1, key2) => {
-        if (object?.[key1] && object?.[key2]) {
-            return `${object[key1]} - ${object[key2]}`;
-        }
-
-        return "--";
-    };
-
-    const mockColumns = [
-        {
-            header: "Prime Award ID",
-            key: "award_id",
-            type: "expandableButton",
-            className: "",
-            element: null
-        },
-        {
-            header: 'Count of Subwards that Match Search Criteria',
-            key: 'subaward_count',
-            type: "alphaNumeric",
-            className: "",
-            element: null
-        },
-        {
-            header: 'Obligations that Match Search Criteria',
-            key: 'subaward_obligation',
-            type: "formatted",
-            className: "",
-            element: (info) => MoneyFormatter.formatMoneyWithPrecision(info.getValue(), 2, "--")
-        }
-    ];
-
-    const columns = useMemo(() => mockColumns.map((col) => {
+    const columns = useMemo(() => columnArray.map((col) => {
         switch (col.type) {
             case "expandableButton":
                 return columnHelper.accessor(col.key, {
@@ -237,8 +51,8 @@ const TanStackTable = (props) => {
                     id: uniqueId(),
                     cell: ({ row, getValue }) => (
                         <button
-                            onClick={() => getSubTable(getValue(), row.id)}
-                            onKeyDown={() => getSubTable(getValue(), row.id)}
+                            onClick={() => togglePrimeAward(getValue(), row.id)}
+                            onKeyDown={() => togglePrimeAward(getValue(), row.id)}
                             role="link"
                             className={`usa-button-link ${col.className ? col.className : ''}`} >
                             <FontAwesomeIcon
@@ -274,169 +88,10 @@ const TanStackTable = (props) => {
                     cell: (info) => info.getValue()
                 });
         }
-    }), [data, expanded]);
-
-    const subColumns = useMemo(() => {
-        if (props.isTransactions) {
-            return [
-                columnHelper.accessor('Award ID', {
-                    header: 'Prime Award ID',
-                    id: uniqueId(),
-                    cell: ({ row, getValue }) => (
-                        <a
-                            target="_blank"
-                            rel="noopener noreferrer"
-                            href={`/award/${row.generated_internal_id}`}
-                            onClick={() => {
-                                props.clickHandler(getValue());
-                            }}>{getValue()}
-                        </a>
-                    )
-                }),
-                columnHelper.accessor('Mod', {
-                    header: 'Modification Number',
-                    id: uniqueId(),
-                    cell: (info) => info.getValue()
-                }),
-                columnHelper.accessor('Recipient Name', {
-                    header: 'Recipient Name',
-                    id: uniqueId(),
-                    cell: (info) => info.getValue()
-                }),
-                columnHelper.accessor('Transaction Amount', {
-                    header: 'Obligations',
-                    id: uniqueId(),
-                    cell: (info) => MoneyFormatter.formatMoneyWithPrecision(info.getValue(), 2, "--")
-                }),
-                columnHelper.accessor('Action Date', {
-                    header: 'Action Date',
-                    id: uniqueId(),
-                    cell: (info) => info.getValue()
-                }),
-                columnHelper.accessor('Transaction Description', {
-                    header: 'Transaction Description',
-                    id: uniqueId(),
-                    cell: (info) => (
-                        <ReadMore
-                            text={info.getValue() || '--'}
-                            limit={90} />
-                    )
-                }),
-                columnHelper.accessor('Action Type', {
-                    header: 'Action Type',
-                    id: uniqueId(),
-                    cell: (info) => info.getValue()
-                }),
-                columnHelper.accessor('Award Type', {
-                    header: 'Award Type',
-                    id: uniqueId(),
-                    cell: (info) => info.getValue()
-                }),
-                columnHelper.accessor('Recipient Location', {
-                    header: 'Recipient Location',
-                    id: uniqueId(),
-                    cell: (info) => pickLocationFormat(info.getValue())
-                }),
-                columnHelper.accessor('Primary Place of Performance', {
-                    header: 'Primary Place of Performance',
-                    id: uniqueId(),
-                    cell: (info) => pickLocationFormat(info.getValue())
-                }),
-                columnHelper.accessor('Awarding Agency', {
-                    header: 'Awarding Agency',
-                    id: uniqueId(),
-                    cell: (info) => info.getValue()
-                }),
-                columnHelper.accessor('Awarding Sub Agency', {
-                    header: 'Awarding Sub Agency',
-                    id: uniqueId(),
-                    cell: (info) => info.getValue()
-                }),
-                columnHelper.accessor('NAICS', {
-                    header: 'North American Industry Classification System (NAICS)',
-                    id: uniqueId(),
-                    cell: (info) => (
-                        <ReadMore
-                            text={twoVariableFormat(info.getValue(), 'code', 'description')}
-                            limit={80} />
-                    )
-                }),
-                columnHelper.accessor('PSC', {
-                    header: 'Product and Service Code (PSC)',
-                    id: uniqueId(),
-                    cell: (info) => (
-                        <ReadMore
-                            text={twoVariableFormat(info.getValue(), 'code', 'description')}
-                            limit={90} />
-                    )
-                })
-            ];
-        }
-
-        return [
-            columnHelper.accessor('Sub-Award ID', {
-                header: 'Subaward ID',
-                id: uniqueId(),
-                cell: ({ row, getValue }) => (
-                    <a
-                        target="_blank"
-                        rel="noopener noreferrer"
-                        href={`/award/${row.prime_award_generated_internal_id}`}
-                        onClick={() => {
-                            props.clickHandler(getValue());
-                        }}>{getValue()}
-                    </a>
-                )
-            }),
-            columnHelper.accessor('Sub-Awardee Name', {
-                header: 'Subrecipient Name',
-                id: uniqueId(),
-                cell: (info) => info.getValue()
-            }),
-            columnHelper.accessor('Sub-Award Amount', {
-                header: 'Subaward Obligations',
-                id: uniqueId(),
-                cell: (info) => MoneyFormatter.formatMoneyWithPrecision(info.getValue(), 2, "--")
-            }),
-            columnHelper.accessor('Sub-Award Date', {
-                header: 'Subaward Action Date',
-                id: uniqueId(),
-                cell: (info) => info.getValue()
-            }),
-            columnHelper.accessor('Sub-Award Description', {
-                header: 'Subaward Description',
-                id: uniqueId(),
-                cell: (info) => (
-                    <ReadMore
-                        text={info.getValue() || '--'}
-                        limit={90} />
-                )
-            }),
-            columnHelper.accessor('Sub-Recipient UEI', {
-                header: 'Subrecipient UEI',
-                id: uniqueId(),
-                cell: (info) => info.getValue()
-            }),
-            columnHelper.accessor('Sub-Recipient Location', {
-                header: 'Subrecipient Location',
-                id: uniqueId(),
-                cell: (info) => pickLocationFormat(info.getValue())
-            }),
-            columnHelper.accessor('Sub-Award Primary Place of Performance', {
-                header: 'Subaward Primary Place of Performance',
-                id: uniqueId(),
-                cell: (info) => pickLocationFormat(info.getValue())
-            }),
-            columnHelper.accessor('Sub-Award Type', {
-                header: 'Subaward Type',
-                id: uniqueId(),
-                cell: (info) => convertToTitleCase(info.getValue())
-            })
-        ];
-    }, [subData, expanded]);
+    }), [props.data, expanded]);
 
     const table = useReactTable({
-        data,
+        data: props.data,
         columns,
         state: {
             expanded
@@ -446,14 +101,6 @@ const TanStackTable = (props) => {
         getFilteredRowModel: getFilteredRowModel(),
         getSortedRowModel: getSortedRowModel(),
         useExpanded
-    });
-
-    const subTable = useReactTable({
-        data: subData,
-        columns: subColumns,
-        getCoreRowModel: getCoreRowModel(),
-        getPaginationRowModel: getPaginationRowModel(),
-        getFilteredRowModel: getFilteredRowModel()
     });
 
     return (
@@ -488,38 +135,10 @@ const TanStackTable = (props) => {
                         {row.getIsExpanded() && (
                             <tr>
                                 <td colSpan={row.getAllCells().length}>
-                                    <table className="usda-table table-for-new-search-page expandable-table">
-                                        <thead>
-                                            {subTable.getHeaderGroups().map((subHeaderGroup) => (
-                                                <tr key={subHeaderGroup.id} className="usda-table__row">
-                                                    {subHeaderGroup.headers.map((subHeader) => (
-                                                        <th key={subHeader.id}>
-                                                            {subHeader.isPlaceholder
-                                                                ? null
-                                                                : flexRender(
-                                                                    subHeader.column.columnDef.header,
-                                                                    subHeader.getContext()
-                                                                )}
-                                                        </th>
-                                                    ))}
-                                                </tr>
-                                            ))}
-                                        </thead>
-                                        <tbody className="usda-table__body">
-                                            {subTable.getRowModel().rows.map((r) => (
-                                                <tr key={r.id}>
-                                                    {r.getVisibleCells().map((subCell) => (
-                                                        <td key={subCell.id}>
-                                                            {flexRender(
-                                                                subCell.column.columnDef.cell,
-                                                                subCell.getContext()
-                                                            )}
-                                                        </td>
-                                                    ))}
-                                                </tr>
-                                            ))}
-                                        </tbody>
-                                    </table>
+
+                                    <NestedTanStackTable
+                                        {...props}
+                                        awardId={awardId} />
                                 </td>
                             </tr>
                         )}
