--- src/js/components/sharedComponents/table/mockData.js
@@ -1,63 +1,0 @@
-
-export const mockRows = [
-    {
-        award_id: "0001",
-        subaward_count: 1510,
-        award_generated_internal_id: "CONT_AWD_0001_9700_FA870215D0001_9700",
-        subaward_obligation: 409303652.12
-    },
-    {
-        award_id: "N0001923C0003",
-        subaward_count: 1723,
-        award_generated_internal_id: "CONT_AWD_N0001923C0003_9700_-NONE-_-NONE-",
-        subaward_obligation: 2406096453.7
-    },
-    {
-        award_id: "N0002416C2229",
-        subaward_count: 1871,
-        award_generated_internal_id: "CONT_AWD_N0002416C2229_9700_-NONE-_-NONE-",
-        subaward_obligation: 552408209.26
-    },
-    {
-        award_id: "N0002415C4301",
-        subaward_count: 1974,
-        award_generated_internal_id: "CONT_AWD_N0002415C4301_9700_-NONE-_-NONE-",
-        subaward_obligation: 658930044.09
-    },
-    {
-        award_id: "N0001921C0020",
-        subaward_count: 2545,
-        award_generated_internal_id: "CONT_AWD_N0001921C0020_9700_-NONE-_-NONE-",
-        subaward_obligation: 492379269.3
-    },
-    {
-        award_id: "W56HZV15C0095",
-        subaward_count: 2676,
-        award_generated_internal_id: "CONT_AWD_W56HZV15C0095_9700_-NONE-_-NONE-",
-        subaward_obligation: 304424719.1
-    },
-    {
-        award_id: "DENA0001942",
-        subaward_count: 2976,
-        award_generated_internal_id: "CONT_AWD_DENA0001942_8900_-NONE-_-NONE-",
-        subaward_obligation: 1572986547.25
-    },
-    {
-        award_id: "N0001920C0009",
-        subaward_count: 4345,
-        award_generated_internal_id: "CONT_AWD_N0001920C0009_9700_-NONE-_-NONE-",
-        subaward_obligation: 419902669.32
-    },
-    {
-        award_id: "N0002418C2106",
-        subaward_count: 12407,
-        award_generated_internal_id: "CONT_AWD_N0002418C2106_9700_-NONE-_-NONE-",
-        subaward_obligation: 918283598.01
-    },
-    {
-        award_id: "N0002416C2116",
-        subaward_count: 45608,
-        award_generated_internal_id: "CONT_AWD_N0002416C2116_9700_-NONE-_-NONE-",
-        subaward_obligation: 2666421624.76
-    }
-];
