+++ src/js/helpers/searchHelper.js
@@ -148,6 +148,12 @@ export const performSpendingBySubawardGrouped = (params) => apiRequest({
     data: params
 });
 
+export const performSpendingByTransactionsGrouped = (params) => apiRequest({
+    url: 'v2/search/spending_by_transaction_grouped',
+    method: 'post',
+    data: params
+});
+
 export const performSubawardSearch = (data) => apiRequest({
     url: 'v2/subawards/',
     method: 'post',
